Break up giant views/database.py into smaller modules
